+++ middleware/favicon/favicon_test.go
@@ -1,6 +1,7 @@
 package favicon
 
 import (
+	"bytes"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -47,13 +48,54 @@ func Test_Middleware_Favicon_Not_Found(t *testing.T) {
 	t.Parallel()
 	defer func() {
 		if err := recover(); err == nil {
-			t.Error("should cache panic")
+			t.Error("should catch panic")
 			return
 		}
 	}()
 
 	fiber.New().Use(New(Config{
 		File: "non-exist.ico",
+	}))
+}
+
+// go test -run Test_Middleware_Favicon_MaxBytes
+func Test_Middleware_Favicon_MaxBytes(t *testing.T) {
+	t.Parallel()
+	defer func() {
+		if err := recover(); err == nil {
+			t.Error("should catch panic")
+		}
+	}()
+
+	dir := t.TempDir()
+	path := dir + "/favicon.ico"
+	err := os.WriteFile(path, bytes.Repeat([]byte("a"), 11), 0o600)
+	require.NoError(t, err)
+
+	fiber.New().Use(New(Config{
+		File:     path,
+		MaxBytes: 10,
+	}))
+}
+
+// go test -run Test_Middleware_Favicon_MaxBytes_FileSystem
+func Test_Middleware_Favicon_MaxBytes_FileSystem(t *testing.T) {
+	t.Parallel()
+	defer func() {
+		if err := recover(); err == nil {
+			t.Error("should catch panic")
+		}
+	}()
+
+	dir := t.TempDir()
+	path := dir + "/favicon.ico"
+	err := os.WriteFile(path, bytes.Repeat([]byte("a"), 11), 0o600)
+	require.NoError(t, err)
+
+	fiber.New().Use(New(Config{
+		File:       "favicon.ico",
+		FileSystem: os.DirFS(dir),
+		MaxBytes:   10,
 	}))
 }